Added more comments, and fixed 2 bugs
Some parentheses were in the wrong place in the "if" statement that calls "getFiles(...)". 
Calloc is now used for "reservedFileNames" so that it is initialized to 0 and more clear how the data is being distributed.
The input and output file pointers now get files opened in main instead of inside of the "getFiles(...)", as they were out of scope when newly set to point to files in a function.