Updated scanner portion
Added getTokenType, worked on scanner portion, output file, and listing file. So far, the scanner can read each character from the input file. It can write to the listing file. But the listing file needs line numbers. Didn't do that yet. There's still quite a bit to do. We need to check and identify each token. The end of the scanner has "if" statements. It should really be switch case statements. But, I haven't worked on that part yet. As far as I know, the pseudo code is enough for us to finish writing the code.